+++ packages/subgraph-service/test/utils/Constants.sol
@@ -12,7 +12,7 @@ abstract contract Constants {
     uint256 internal constant minimumProvisionTokens = 1000 ether;
     uint256 internal constant maximumProvisionTokens = type(uint256).max;
     uint32 internal constant delegationRatio = 16;
-    uint32 public constant stakeToFeesRatio = 2;
+    uint256 public constant stakeToFeesRatio = 2;
     uint256 public constant maxPOIStaleness = 28 days;
     uint128 public constant serviceCut = 10000; 
     uint128 public constant curationCut = 10000;
